backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-multipart
ezdxf
shapely
networkx
matplotlib
google-generativeai
pytest
httpx
